requirements.txt
# Existing requirements
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
asyncpg>=0.28.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
aiohttp>=3.8.0
requests>=2.31.0
plotly>=5.15.0
matplotlib>=3.7.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
websockets>=11.0.0
numpy>=1.24.0
python-dateutil>=2.8.0

# New requirements for interface
python-multipart>=0.0.6
jinja2>=3.1.0
flask
plotly
pandas